requirements.txt
astral==2.2
ws4py==0.5.1
requests==2.20.0
paho-mqtt==1.5.0
email-validator==1.1.1
flask==1.1.2
flask_login==0.5.0
flask_restful==0.3.8
flask_wtf==0.14.3
jsonify==0.5
Werkzeug==1.0.1
WTForms==2.3.3
